@@ -4,7 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at http://mozilla.org/MPL/2.0/.
 
-// +build noquic !go1.14 go1.17
+//go:build noquic || !go1.15
+// +build noquic !go1.15
 
 package connections
 
