@@ -4,7 +4,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-//+build !windows
+//go:build !windows
+// +build !windows
 
 // (No syscall.Umask or the equivalent on Windows)
 
