@@ -4,7 +4,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this file,
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
-// Commmand stfindignored lists ignored files under a given folder root.
+// Command stfindignored lists ignored files under a given folder root.
 package main
 
 import (
